Use usize instead of u64 for binary storage interface

+++ src/storage/binary_storage.rs
@@ -33,63 +33,63 @@ pub trait BinaryStorage {
 
   fn is_open(&self) -> bool;
 
-  fn w_i8(&mut self, offset: u64, data: i8) -> Result<(), Error>;
-  fn w_i16(&mut self, offset: u64, data: i16) -> Result<(), Error>;
-  fn w_i32(&mut self, offset: u64, data: i32) -> Result<(), Error>;
-  fn w_i64(&mut self, offset: u64, data: i64) -> Result<(), Error>;
+  fn w_i8(&mut self, offset: usize, data: i8) -> Result<(), Error>;
+  fn w_i16(&mut self, offset: usize, data: i16) -> Result<(), Error>;
+  fn w_i32(&mut self, offset: usize, data: i32) -> Result<(), Error>;
+  fn w_i64(&mut self, offset: usize, data: i64) -> Result<(), Error>;
 
-  fn w_u8(&mut self, offset: u64, data: u8) -> Result<(), Error>;
-  fn w_u16(&mut self, offset: u64, data: u16) -> Result<(), Error>;
-  fn w_u32(&mut self, offset: u64, data: u32) -> Result<(), Error>;
-  fn w_u64(&mut self, offset: u64, data: u64) -> Result<(), Error>;
+  fn w_u8(&mut self, offset: usize, data: u8) -> Result<(), Error>;
+  fn w_u16(&mut self, offset: usize, data: u16) -> Result<(), Error>;
+  fn w_u32(&mut self, offset: usize, data: u32) -> Result<(), Error>;
+  fn w_u64(&mut self, offset: usize, data: u64) -> Result<(), Error>;
 
-  fn w_f32(&mut self, offset: u64, data: f32) -> Result<(), Error>;
-  fn w_f64(&mut self, offset: u64, data: f64) -> Result<(), Error>;
+  fn w_f32(&mut self, offset: usize, data: f32) -> Result<(), Error>;
+  fn w_f64(&mut self, offset: usize, data: f64) -> Result<(), Error>;
 
-  fn w_bool(&mut self, offset: u64, data: bool) -> Result<(), Error>;
+  fn w_bool(&mut self, offset: usize, data: bool) -> Result<(), Error>;
 
-  fn w_bytes(&mut self, offset: u64, data: &[u8]) -> Result<(), Error>;
-  fn w_str(&mut self, offset: u64, data: &str) -> Result<(), Error>;
+  fn w_bytes(&mut self, offset: usize, data: &[u8]) -> Result<(), Error>;
+  fn w_str(&mut self, offset: usize, data: &str) -> Result<(), Error>;
 
 
-  fn r_i8(&self, offset: u64) -> Result<i8, Error>;
-  fn r_i16(&self, offset: u64) -> Result<i16, Error>;
-  fn r_i32(&self, offset: u64) -> Result<i32, Error>;
-  fn r_i64(&self, offset: u64) -> Result<i64, Error>;
+  fn r_i8(&self, offset: usize) -> Result<i8, Error>;
+  fn r_i16(&self, offset: usize) -> Result<i16, Error>;
+  fn r_i32(&self, offset: usize) -> Result<i32, Error>;
+  fn r_i64(&self, offset: usize) -> Result<i64, Error>;
 
-  fn r_u8(&self, offset: u64) -> Result<u8, Error>;
-  fn r_u16(&self, offset: u64) -> Result<u16, Error>;
-  fn r_u32(&self, offset: u64) -> Result<u32, Error>;
-  fn r_u64(&self, offset: u64) -> Result<u64, Error>;
+  fn r_u8(&self, offset: usize) -> Result<u8, Error>;
+  fn r_u16(&self, offset: usize) -> Result<u16, Error>;
+  fn r_u32(&self, offset: usize) -> Result<u32, Error>;
+  fn r_u64(&self, offset: usize) -> Result<u64, Error>;
 
-  fn r_f32(&self, offset: u64) -> Result<f32, Error>;
-  fn r_f64(&self, offset: u64) -> Result<f64, Error>;
+  fn r_f32(&self, offset: usize) -> Result<f32, Error>;
+  fn r_f64(&self, offset: usize) -> Result<f64, Error>;
 
-  fn r_bool(&self, offset: u64) -> Result<bool, Error>;
+  fn r_bool(&self, offset: usize) -> Result<bool, Error>;
 
-  fn r_bytes(&self, offset: u64, len: usize) -> Result<Vec<u8>, Error>;
-  fn r_str(&self, offset: u64, len: usize) -> Result<String, Error>;
+  fn r_bytes(&self, offset: usize, len: usize) -> Result<Vec<u8>, Error>;
+  fn r_str(&self, offset: usize, len: usize) -> Result<String, Error>;
 
   fn fill(
     &mut self, 
-    start: Option<u64>, 
-    end: Option<u64>, 
+    start: Option<usize>, 
+    end: Option<usize>, 
     val: u8
   ) -> Result<(), Error>;
 
   fn is_filled(
     &self, 
-    start: Option<u64>, 
-    end: Option<u64>, 
+    start: Option<usize>, 
+    end: Option<usize>, 
     val: u8
   ) -> Result<bool, Error>;
 
-  fn get_expand_size(&self) -> u64;
-  fn set_expand_size(&mut self, expand_size: u64) -> Result<(), Error>;
+  fn get_expand_size(&self) -> usize;
+  fn set_expand_size(&mut self, expand_size: usize) -> Result<(), Error>;
 
-  fn get_capacity(&self) -> Result<u64, Error>;
+  fn get_capacity(&self) -> Result<usize, Error>;
 
-  fn expand(&mut self, min_capacity: u64) -> Result<(), Error>;
+  fn expand(&mut self, min_capacity: usize) -> Result<(), Error>;
 
 }
